Updated Code

+++ phitb_product/grails-app/services/phitb_product/DivisionGroupRegisterService.groovy
@@ -28,7 +28,7 @@ class DivisionGroupRegisterService {
         return DivisionGroupRegister.findById(Long.parseLong(id))
     }
 
-    def getAllByEntity(long limit, long offset, long entityId)
+    def getAllByEntity(int limit, int offset, long entityId)
     {
         Integer o = offset ? Integer.parseInt(offset.toString()) : 0
         Integer l = limit ? Integer.parseInt(limit.toString()) : 100
